Split createCurl function into 2 parts for both Puser and non-Pusher endpoints
For the upcoming attachment upload change, we need to use the Curl interface with S3 as well as the Pusher API. Therefore this re-uses the same Curl setup logic for 'raw' requests, and moves the Pusher-specific logic to a wrapper around that.

Additionally, we now send the Content-Length header on all requests.